+++ src/main/java/com/app/housing_association/security/model/LoggedUser.java
@@ -1,4 +1,4 @@
-package com.app.housing_association.security.login.model;
+package com.app.housing_association.security.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
