+++ src/main/java/com/app/housing_association/security/utils/JwtUtils.java
@@ -1,6 +1,6 @@
-package com.app.housing_association.security.login.utils;
+package com.app.housing_association.security.utils;
 
-import com.app.housing_association.security.login.model.AppUserDetails;
+import com.app.housing_association.security.model.AppUserDetails;
 import io.jsonwebtoken.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
